moved video stuff to src/video
