+++ src/video/camera.h
@@ -22,7 +22,7 @@
 #include <QMutex>
 #include "vpx/vpx_image.h"
 #include "opencv2/opencv.hpp"
-#include "src/videosource.h"
+#include "src/video/videosource.h"
 
 class CameraWorker;
