+++ src/video/cameraworker.cpp
