+++ src/video/cameraworker.h
