+++ src/video/netvideosource.h
