+++ src/video/videosource.h
